+++ test/runner/browser/chrome_test.dart
@@ -13,6 +13,7 @@ import 'package:shelf/shelf.dart' as shelf;
 import 'package:shelf/shelf_io.dart' as shelf_io;
 import 'package:shelf_web_socket/shelf_web_socket.dart';
 
+import '../../io.dart';
 import '../../utils.dart';
 
 void main() {
@@ -140,6 +141,6 @@ webSocket.addEventListener("open", function() {
     var chrome = new Chrome("http://dart-lang.org",
         executable: "_does_not_exist");
     expect(chrome.onExit, throwsA(isApplicationException(startsWith(
-        "Failed to start Chrome: No such file or directory"))));
+        "Failed to start Chrome: $noSuchFileMessage"))));
   });
 }