+++ test/runner/browser/dartium_test.dart
@@ -150,7 +150,7 @@ webSocket.onOpen.first.then((_) =>
     var dartium = new Dartium("http://dart-lang.org",
         executable: "_does_not_exist");
     expect(dartium.onExit, throwsA(isApplicationException(startsWith(
-        "Failed to start Dartium: No such file or directory"))));
+        "Failed to start Dartium: $noSuchFileMessage"))));
   });
 
   test("can run successful tests", () {