+++ test/runner/browser/firefox_test.dart
@@ -150,7 +150,7 @@ webSocket.addEventListener("open", function() {
     var firefox = new Firefox("http://dart-lang.org",
         executable: "_does_not_exist");
     expect(firefox.onExit, throwsA(isApplicationException(startsWith(
-        "Failed to start Firefox: No such file or directory"))));
+        "Failed to start Firefox: $noSuchFileMessage"))));
   });
 
   group("can run successful tests", () {