+++ test/runner/loader_test.dart
@@ -90,7 +90,8 @@ void main() {
     test("throws a nice error if the package root doesn't exist", () {
       expect(() => new Loader([TestPlatform.chrome], root: _sandbox),
           throwsA(isApplicationException(
-              "Directory ${p.join(_sandbox, 'packages')} does not exist.")));
+              "Directory ${p.prettyUri(p.toUri(p.join(_sandbox, 'packages')))} "
+                  "does not exist.")));
     });
   });
 
@@ -132,7 +133,7 @@ void main() {
         expect(suites.map((suite) => suite.path), unorderedEquals([
           p.join(_sandbox, 'a_test.dart'),
           p.join(_sandbox, 'another_test.dart'),
-          p.join(_sandbox, 'dir/sub_test.dart')
+          p.join(_sandbox, 'dir', 'sub_test.dart')
         ]));
       });