+++ test/runner/signal_test.dart
@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@TestOn("vm")
+// Windows doesn't support sending signals.
+@TestOn("vm && !windows")
 
 import 'dart:async';
 import 'dart:convert';