+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * When running a test suite via `dart path/to/test.dart`, throw an exception if
   the suite fails so that the exit code is set properly.
 
+* Add support for running on Windows.
+
 ### 0.12.0-beta.10
 
 * Fix running browser tests in subdirectories.